pyproject.toml
[project]
name = "test-hitalent"
version = "0.1.0"
description = ""
authors = [
    {name = "TheAppleKingy",email = "pridanov.ali.job@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.2,<2.0.0)",
    "dishka (>=1.7.2,<2.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
